+++ client/src/pages/screens/earnBattery/EarnBattery.jsx
@@ -78,6 +78,11 @@ const EarnBattery = ({ user, globalLoading, updateGlobalValues }) => {
   const userHasStarred = globalLoading ? false : user.earnBattery.githubStar;
   const userHasSubscribed = globalLoading ? false : !!user.account.subscription;
 
+  const getCode = (game)=> {
+    if (!user || globalLoading)return
+    return user.earnBattery[game];
+  }
+
   const getButton = (game) => {
     if (!user || globalLoading) return;
     let onclick = () => console.log("Click..");
@@ -86,7 +91,7 @@ const EarnBattery = ({ user, globalLoading, updateGlobalValues }) => {
     let cursor = "pointer";
     let width = "100%";
 
-    const currentGame = user.earnBattery[game];
+    const currentGame = getCode(game)
     if (currentGame) {
       onclick = () => navigator.clipboard.writeText(currentGame);
       cursor = "copy";
@@ -261,7 +266,9 @@ const EarnBattery = ({ user, globalLoading, updateGlobalValues }) => {
             <a target="_blank" rel="noopener noreferrer" href={URLS.megarpg}>
               MEGA rpg
             </a>{" "}
-            Discord server and type !hack 'code'{" "}
+            Discord server and type in{" "}
+            <strong>!hack {getCode("megarpg") || "CODE"}</strong> in the
+            play-here channel!
           </CardText>
           {getButton("megarpg")}
         </CardBody>
@@ -287,7 +294,8 @@ const EarnBattery = ({ user, globalLoading, updateGlobalValues }) => {
             <a target="_blank" rel="noopener noreferrer" href={URLS.chessathor}>
               Chessathor
             </a>{" "}
-            and input the 'code' to receive the battery bonus!{" "}
+            and input <strong>{getCode("chessathor") || "the personal code"}</strong> as a highscore
+            name to receive the battery bonus!
           </CardText>
           {getButton("chessathor")}
         </CardBody>